Functions finished

+++ public/assets/js/crudFunctions.js
@@ -2,35 +2,42 @@ const { remote} = require('electron')
 const main = remote.require('./../src/main.js')
 
 //Validar fecha para que sea de lunes a viernes y la hora de 8 a 7 
-function agendarReservacion(idPaquete){
+async function agendarReservacion(idPaquete){
     var cliente = document.getElementById('cliente');
     var fecha_contratacion = document.getElementById('fecha_contratacion');
     var fecha_finalizacion = document.getElementById('fecha_finalizacion');
     var periodo_uso = document.getElementById('periodo_uso');
     var tiempo_uso = document.getElementById('tiempo_uso');
     var total_neto = document.getElementById('total_neto');
+    var disponibilidad = await main.verificarDisponibilidad(fecha_contratacion.value, idPaquete);
+    console.log(disponibilidad);
+    if(disponibilidad == 0){
+        var today = new Date();
+        var dd = String(today.getDate()).padStart(2, '0');
+        var mm = String(today.getMonth() + 1).padStart(2, '0'); //Month starts in 0
+        var yyyy = today.getFullYear();
 
-    var today = new Date();
-    var dd = String(today.getDate()).padStart(2, '0');
-    var mm = String(today.getMonth() + 1).padStart(2, '0'); //Month starts in 0
-    var yyyy = today.getFullYear();
+        today = yyyy + '/' + mm + '/' + dd;
 
-    today = yyyy + '/' + mm + '/' + dd;
-    //document.write(today);
-
-    const newReservacion = {
-        idPaquete: idPaquete,
-        cliente: cliente.value,
-        fecha_registro: today,
-        fecha_contratacion: fecha_contratacion.value,
-        fecha_finalizacion: fecha_finalizacion.value,
-        periodo_uso: periodo_uso.value,
-        tiempo_uso: tiempo_uso.value,
-        total_neto: total_neto.value,
-        pagado: 0
+        const newReservacion = {
+            idPaquete: idPaquete,
+            cliente: cliente.value,
+            fecha_registro: today,
+            fecha_contratacion: fecha_contratacion.value,
+            fecha_finalizacion: fecha_finalizacion.value,
+            periodo_uso: periodo_uso.value,
+            tiempo_uso: tiempo_uso.value,
+            total_neto: total_neto.value,
+            pagado: 0
+        }
+        main.createReservacion(newReservacion, today);
+        location.reload(); 
     }
-    main.createReservacion(newReservacion, today);
-    location.reload();  
+    else{
+        alert("El paquete solicitado no está disponible para la fecha que deseas, revisa el calendario de reservaciones y selecciona una fecha diferente");
+    }
+     
+    //Show a notification when reservation added
 }
 
 function renderPaquetes(paquetes){
@@ -125,17 +132,67 @@ function renderNoPagados(clientesNP){
     tablaHTML.innerHTML = tabla;
 }
 
-async function calcularPrecios(){
+function futureDate(fecha_contratacion, cantDias, cantHoras){
+    var fecha_actual = new Date(fecha_contratacion.value);
+    fecha_actual.setDate(fecha_actual.getUTCDate()+parseInt(cantDias));   
+    fecha_actual.setHours(fecha_actual.getHours()+parseInt(cantHoras));
+    var futureDate = fecha_actual.getFullYear()+'-'+('0'+(fecha_actual.getMonth()+1)).slice(-2)+'-'+('0'+(fecha_actual.getDate())).slice(-2) + "T"+("0"+fecha_actual.getHours()).slice(-2)+":"+("0"+fecha_actual.getMinutes()).slice(-2);;
+    return futureDate;
+}
+
+async function calcularFechas(fecha_contratacion, fecha_finalizacion, periodo_uso, tiempo_uso){
+    var newFecha_finalizacion;
+    console.log(periodo_uso);
+    switch(periodo_uso){
+        case '1':
+            newFecha_finalizacion = futureDate(fecha_contratacion, 0, tiempo_uso);
+            break;
+        case '2':
+            var fecha = new Date(fecha_contratacion.value);
+            var cambio = 5-fecha.getUTCDay();
+            console.log(cambio);
+            if(cambio<tiempo_uso){
+                newFecha_finalizacion = futureDate(fecha_contratacion, 1*tiempo_uso+2, 0);
+            }
+            else{
+                newFecha_finalizacion = futureDate(fecha_contratacion, 1*tiempo_uso, 0);
+            }
+            break;
+        case '3':
+            newFecha_finalizacion = futureDate(fecha_contratacion, 7*tiempo_uso, 0);
+            break;
+        case '4':
+            newFecha_finalizacion = futureDate(fecha_contratacion, 30*tiempo_uso, 0);
+            break;
+        default:
+            console.log("no entro a ningun case")
+            break;
+    }
+    fecha_finalizacion.value = newFecha_finalizacion;
+}
+
+async function calcular(){
     var idPaquete = document.getElementById('idpaquete').value;
     var periodo = document.getElementById('periodo_uso').value;
-    var tiempo = document.getElementById('tiempo_uso').value;
+    var tiempo = document.getElementById('tiempo_uso');
+    var fecha_contratacion = document.getElementById('fecha_contratacion');
+    var fecha_finalizacion = document.getElementById('fecha_finalizacion');
     const total_neto = document.getElementById('total_neto');
     const total_iva = document.getElementById('total_iva');
-    
+
+    if(periodo == 1){
+        tiempo.setAttribute('max', '2'); 
+    }
+    else{
+        tiempo.removeAttribute('max');
+    }
+
+    calcularFechas(fecha_contratacion, fecha_finalizacion, periodo, tiempo.value);
+
     paquete = await main.getPaquete(idPaquete);
     let preciosPeriodo = [paquete['tarifa_hora'], paquete['tarifa_dia'], paquete['tarifa_semana'], paquete['tarifa_mensual']];
     
-    costo_neto=parseFloat(preciosPeriodo[periodo-1]*tiempo);
+    costo_neto=parseFloat(preciosPeriodo[periodo-1]*tiempo.value);
 
     total_neto.value = Math.round((costo_neto + Number.EPSILON) * 100) / 100 ;
     total_iva.value = Math.round(((costo_neto*1.16) + Number.EPSILON) * 100) / 100 ;
@@ -147,8 +204,11 @@ async function getNoPagados(){
 }
 
 async function deleteReservacion(id){
-    await main.deleteReservacion(id);
-    location.reload();
+    const response = confirm("Seguro que deseas eliminar esta reservacion?");
+    if(response){
+        await main.deleteReservacion(id);
+        location.reload();
+    }
 }
 
 async function validateReservacion(id){
@@ -159,14 +219,14 @@ async function validateReservacion(id){
 async function getPaquetes() {
     paquetes = await main.listPaquetes();
     renderPaquetes(paquetes);
-};
+}
 
 async function fillModalReservacion(id) {
     paquete = await main.getPaquete(id);
     var formAgendar = document.getElementById('reservacionForm');
     var idPaquete = document.getElementById('idpaquete');
     var titulo = document.getElementById("titleReservacion");
-
+    
     formAgendar.reset();
     idPaquete.value = paquete.idpaquete;
     titulo.innerHTML = '';
@@ -199,7 +259,6 @@ async function fillModalModificar(id){
     tarifa_mensual.value = paquete.tarifa_mensual;
     formModificar.setAttribute('onsubmit', 'modificarPaquete('+paquete.idpaquete+')'); //Hacer funcion para modificar paquete
 }
-
  
 async function modificarPaquete(id){
     var nombre = document.getElementById('modificarnombre').value;
